requirements.txt
amqp==1.4.6
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
blessings==1.6
bpython==0.14.1
celery==3.1.17
certifi==14.05.14
cffi==0.9.2
curtsies==0.1.19
Django==1.9
django-oidc-provider==0.2.5
django-recaptcha==1.0.4
django-redis==3.8.3
django-redis-sessions==0.4.0
enum34==1.0.4
flower==0.7.3
greenlet==0.4.5
httplib2==0.9
Jinja2==2.7.3
kombu==3.0.24
MarkupSafe==0.23
ndg-httpsclient==0.3.3
oauthlib==0.7.2
psycopg2==2.5.4
pyasn1==0.1.7
pycparser==2.10
Pygments==2.0.2
PyJWT==1.1.0
pyOpenSSL==0.14
python-dateutil==2.4.0
python-openid==2.2.5
python-social-auth==0.2.6
pytz==2014.10
raven==5.3.1
redis==2.10.3
requests
requests-oauthlib==0.4.2
simplejson==3.6.5
six==1.9.0
tornado==4.1
xmltodict==0.9.2
rollbar
markdown
django-filter